Added chord variation authorization

+++ backend/Chordy/WebApi/Controllers/ChordVariationsController.cs
@@ -1,5 +1,6 @@
 ﻿using Chordy.BusinessLogic.Interfaces;
 using Chordy.BusinessLogic.Models;
+using Chordy.BusinessLogic.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -7,13 +8,17 @@ namespace Chordy.WebApi.Controllers
 {
     [ApiController]
     [Route("api/chord-variations")]
-    public class ChordVariationController(IChordVariationService chordVariationService) : ControllerBase
+    public class ChordVariationController(IChordVariationService chordVariationService, IAuthorizationService authorizationService) : ControllerBase
     {
         [HttpPost]
         [Authorize]
         public async Task<IActionResult> CreateAsync([FromBody] ChordVariationCreateDto dto, CancellationToken cancellationToken)
         {
-            var chordVariation = await chordVariationService.AddAsync(dto, cancellationToken);
+            var userIdClaim = User.FindFirst("userId")?.Value;
+            if (!Guid.TryParse(userIdClaim, out var userId))
+                return Unauthorized();
+
+            var chordVariation = await chordVariationService.AddAsync(dto, userId, cancellationToken);
             return CreatedAtAction("GetById", new { id = chordVariation.Id }, chordVariation);
         }
 
@@ -31,18 +36,35 @@ namespace Chordy.WebApi.Controllers
             return Ok(variation);
         }
 
+        [HttpGet]
+        public async Task<IActionResult> GetAllAsync(CancellationToken cancellationToken)
+        {
+            var chordVariations = await chordVariationService.GetAllAsync(cancellationToken);
+            return Ok(chordVariations);
+        }
+
         [HttpPut("{id:int}")]
-        [Authorize(Roles = "admin")]
+        [Authorize]
         public async Task<IActionResult> UpdateAsync([FromRoute] int id, [FromBody] ChordVariationCreateDto dto, CancellationToken cancellationToken)
         {
+            var authorizationResult = await authorizationService.AuthorizeAsync(User, id, "ChordVariationOwnerOrAdmin");
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
             await chordVariationService.UpdateAsync(id, dto, cancellationToken);
             return NoContent();
         }
 
         [HttpDelete("{id:int}")]
-        [Authorize(Roles = "admin")]
+        [Authorize]
         public async Task<IActionResult> DeleteAsync([FromRoute] int id, CancellationToken cancellationToken)
         {
+            var authorizationResult = await authorizationService.AuthorizeAsync(User, id, "ChordVariationOwnerOrAdmin");
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
             await chordVariationService.DeleteAsync(id, cancellationToken);
             return NoContent();
         }
